@@ -36,18 +36,19 @@
   "dependencies": {
     "@prisma/adapter-pg": "^6.18.0",
     "@prisma/client": "^6.18.0",
+    "@prisma/extension-accelerate": "^2.0.2",
     "@sentry/nextjs": "^10.22.0",
     "dotenv": "^17.2.3",
     "next": "16.0.0",
-    "next-devtools-mcp": "^0.2.1",
+    "next-devtools-mcp": "^0.2.2",
     "prisma": "^6.18.0",
     "react": "19.2.0",
     "react-dom": "19.2.0"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.1.1",
+    "@eslint/compat": "^1.4.0",
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.8.0",
+    "@eslint/js": "^9.38.0",
     "@next/eslint-plugin-next": "16.0.0",
     "@nx/devkit": "22.1.0-beta.0",
     "@nx/eslint": "22.1.0-beta.0",
@@ -56,19 +57,19 @@
     "@nx/next": "22.1.0-beta.0",
     "@nx/playwright": "22.1.0-beta.0",
     "@nx/workspace": "22.1.0-beta.0",
-    "@playwright/test": "^1.36.0",
+    "@playwright/test": "^1.56.1",
     "@swc-node/register": "~1.11.1",
     "@swc/cli": "~0.7.8",
     "@swc/core": "~1.13.5",
-    "@swc/helpers": "~0.5.11",
+    "@swc/helpers": "~0.5.17",
     "@tailwindcss/postcss": "^4.1.16",
     "@tanstack/react-query-devtools": "^5.90.2",
     "@types/react": "^19.2.2",
     "@types/react-dom": "^19.2.2",
     "autoprefixer": "10.4.21",
-    "eslint": "^9.8.0",
+    "eslint": "^9.38.0",
     "eslint-config-next": "16.0.0",
-    "eslint-config-prettier": "^10.0.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-import": "2.32.0",
     "eslint-plugin-jsx-a11y": "6.10.2",
     "eslint-plugin-playwright": "^2.2.2",
@@ -81,9 +82,9 @@
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.7.1",
     "tailwindcss": "4.1.16",
-    "tslib": "^2.3.0",
-    "typescript": "~5.9.2",
-    "typescript-eslint": "^8.40.0"
+    "tslib": "^2.8.1",
+    "typescript": "~5.9.3",
+    "typescript-eslint": "^8.46.2"
   },
   "workspaces": [
     "apps/*",
