@@ -1,15 +1,10 @@
-const { join } = require('path');
-
 // Note: If you use library-specific PostCSS/Tailwind configuration then you should remove the `postcssConfig` build
 // option from your application's configuration (i.e. project.json).
 //
 // See: https://nx.dev/guides/using-tailwind-css-in-react#step-4:-applying-configuration-to-libraries
 
-module.exports = {
+export default {
   plugins: {
-    '@tailwindcss/postcss': {
-      config: join(__dirname, 'tailwind.config.ts'),
-    },
-    autoprefixer: {},
+    '@tailwindcss/postcss': {},
   },
 };
