@@ -1,15 +1,15 @@
 import { PrismaClient } from './generated/client';
+import { withAccelerate } from '@prisma/extension-accelerate';
 
 const globalForPrisma = globalThis as unknown as {
-  prisma?: PrismaClient;
+  prisma?: ReturnType<typeof createPrismaClient>;
 };
 
 const isDev = process.env['NODE_ENV'] === 'development';
 const enableQueryLog = process.env['PRISMA_QUERY_LOG'] === 'true' || isDev;
 
-export const prisma =
-  globalForPrisma.prisma ??
-  new PrismaClient({
+function createPrismaClient() {
+  return new PrismaClient({
     log: enableQueryLog
       ? [
           { emit: 'event', level: 'query' },
@@ -20,21 +20,29 @@ export const prisma =
           { emit: 'stdout', level: 'error' },
           { emit: 'stdout', level: 'warn' },
         ],
-  });
+  }).$extends(withAccelerate());
+}
+
+export const prisma = globalForPrisma.prisma ?? createPrismaClient();
 
 if (enableQueryLog && isDev) {
-  // Type assertion needed due to conditional log configuration
-  const queryableClient = prisma as PrismaClient & {
-    $on: (
-      event: 'query',
-      callback: (e: { query: string; params: string; duration: number }) => void
-    ) => void;
-  };
-  queryableClient.$on('query', (e: { query: string; params: string; duration: number }) => {
-    console.warn('Query: ' + e.query);
-    console.warn('Params: ' + e.params);
-    console.warn('Duration: ' + e.duration + 'ms');
+  // Access the underlying client for query event logging
+  const baseClient = new PrismaClient({
+    log: [
+      { emit: 'event', level: 'query' },
+      { emit: 'stdout', level: 'error' },
+      { emit: 'stdout', level: 'warn' },
+    ],
   });
+
+  baseClient.$on(
+    'query',
+    (e: { query: string; params: string; duration: number }) => {
+      console.warn('Query: ' + e.query);
+      console.warn('Params: ' + e.params);
+      console.warn('Duration: ' + e.duration + 'ms');
+    }
+  );
 }
 
 if (process.env['NODE_ENV'] !== 'production') {
