@@ -305,7 +305,7 @@ const config = {
       "value": "prisma-client-js"
     },
     "output": {
-      "value": "/home/runner/work/ws-ops/ws-ops/packages/database/src/generated/client",
+      "value": "D:\\GitHub\\ws-ops\\packages\\database\\src\\generated\\client",
       "fromEnvVar": null
     },
     "config": {
@@ -314,12 +314,12 @@ const config = {
     "binaryTargets": [
       {
         "fromEnvVar": null,
-        "value": "debian-openssl-3.0.x",
+        "value": "windows",
         "native": true
       }
     ],
     "previewFeatures": [],
-    "sourceFilePath": "/home/runner/work/ws-ops/ws-ops/packages/database/prisma/schema.prisma",
+    "sourceFilePath": "D:\\GitHub\\ws-ops\\packages\\database\\prisma\\schema.prisma",
     "isCustomOutput": true
   },
   "relativeEnvPaths": {
@@ -332,8 +332,7 @@ const config = {
     "db"
   ],
   "activeProvider": "postgresql",
-  "postinstall": true,
-  "ciName": "GitHub Actions",
+  "postinstall": false,
   "inlineDatasources": {
     "db": {
       "url": {
@@ -352,8 +351,8 @@ const fs = require('fs')
 config.dirname = __dirname
 if (!fs.existsSync(path.join(__dirname, 'schema.prisma'))) {
   const alternativePaths = [
-    "packages/database/src/generated/client",
-    "database/src/generated/client",
+    "src/generated/client",
+    "generated/client",
   ]
   
   const alternativePath = alternativePaths.find((altPath) => {
@@ -382,8 +381,8 @@ exports.PrismaClient = PrismaClient
 Object.assign(exports, Prisma)
 
 // file annotations for bundling tools to include these files
-path.join(__dirname, "libquery_engine-debian-openssl-3.0.x.so.node");
-path.join(process.cwd(), "packages/database/src/generated/client/libquery_engine-debian-openssl-3.0.x.so.node")
+path.join(__dirname, "query_engine-windows.dll.node");
+path.join(process.cwd(), "src/generated/client/query_engine-windows.dll.node")
 // file annotations for bundling tools to include these files
 path.join(__dirname, "schema.prisma");
-path.join(process.cwd(), "packages/database/src/generated/client/schema.prisma")
+path.join(process.cwd(), "src/generated/client/schema.prisma")
